refactoring redux buttons
